requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
captcha==0.6.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pillow==11.0.0
propcache==0.2.0
python-dotenv==1.0.1
yarl==1.15.4
